backend/ocr-service/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
paddleocr==2.7.0.3
paddlepaddle==2.6.2
pillow==10.4.0
numpy==1.26.4
requests==2.32.3
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
